handle multiple messages in one telnet read
Sometimes two messages were in the telnet buffer, and removing the \r\n
caused them to be handed to parse() as one message. It then could not
validate the structure, as the total length mismatched the first given
length header.

This fixes this by splitting the messages and parsing them separately
from each other.